+++ _randomgen/randomgen/__init__.py
@@ -8,9 +8,12 @@ from randomgen.threefry import ThreeFry
 from randomgen.threefry32 import ThreeFry32
 from randomgen.xoroshiro128 import Xoroshiro128
 from randomgen.xorshift1024 import Xorshift1024
+from randomgen.xoshiro256starstar import Xoshiro256StarStar
+from randomgen.xoshiro512starstar import Xoshiro512StarStar
 
 __all__ = ['RandomGenerator', 'DSFMT', 'MT19937', 'PCG64', 'PCG32', 'Philox',
-           'ThreeFry', 'ThreeFry32', 'Xoroshiro128', 'Xorshift1024']
+           'ThreeFry', 'ThreeFry32', 'Xoroshiro128', 'Xorshift1024',
+           'Xoshiro256StarStar', 'Xoshiro512StarStar']
 
 from ._version import get_versions