+++ _randomgen/randomgen/src/xoshiro512starstar/xoshiro512starstar.orig.h
@@ -1,0 +1,6 @@
+#include <stdint.h>
+#include <string.h>
+
+uint64_t s[8];
+uint64_t next(void);
+void jump(void);