+++ _randomgen/doc/source/change-log.rst
@@ -1,6 +1,13 @@
 Change Log
 ----------
 
+After v1.15
+===========
+- Added Xoshiro256** and Xoshiro512**, the preferred generators of this class
+- Fixed bug in `jump` method of Random123 generators which did nto specify a default value
+
+
+
 v1.15
 =====
 - Synced empty choice changes